fix(Notes Widget): avoid HTTP 500 due mid-UTF8 excerpt truncation
Fix courtesty due to @D-side

This issue occurs when 100 bytes past the start of the note is inside
a UTF-8 multibyte character. E. g. if a note is comprised of 99
single-bytes and 100th multibyte.

This is very easy to hit in Russian, since it uses a Cyrillic charset
(2 bytes per UTF-8 char), but spaces and punctuation from ASCII
(1 byte each). But since the condition is so obscure, some texts
work and some don't seemingly at random (except not!).

Based on https://stackoverflow.com/q/9087502/ it seems that
using  `mb_substr` with `utf-8` seems to be the better approach

Signed-off-by: Benjamin Schmid <bentolor@users.noreply.github.com>